Fully implemented audio module and changed loot
Notes:
Loot now has a %chance not to spawn.
Renamed the suspense sound folder.
All sounds need to be louder.
Beach sounds needed.
The audio module currently uses a slightly hacky method for suspense
sounds - when the next version of CC is release this'll be updated.
The localized sounds need to be made simpler, as their current
implementation is overly complex.