Minor fix for small error in wounds.
Note, forgot to mention the following in previous commit:
Fixed a major bug where transitioning with wounds wound screw up the activity and require it to be restarted.
Made wounds' positions save on transition, so the re-added wounds are placed where they should. Need to determine whether or not emission state should be saved as well.